Add most tests + updates
Move Snapshot class used in tests to it's own module. Makes it cleaner and easier to control.

Add Repository tests for findById and save, conflict strategy tests are marked as TODO.

Added some TODO comments on what to update in types.

Add the checkConcurrencyStrategy logic.

Update the Partition classes used for testing.