Update jest config, Repository and package
Add compile and compile.watch scripts

Add main and types to package.json

Update jest config to disable type checking by ts-jest. It seems to give a lot of false positives when passing subclasses as arguments. Such as "Location is not assignable to Aggregate<BaseState>". Yes, ts-jest. It is. We get type checking from the compilation step and IDE help anyway.