Class syntax, move test files, rename files, use import/export
Refactor to use class syntax instead of adding to prototype.

Use import/export instead of require & module.exports

Break out fixtures to __fixtures__ dirs.

Move test suites to be siblings to the module being tested.

Break out partition fixtures to own classes. Makes it easier to understand and extend if needed.

Rename files to be named as the class in PascalCase instead of snake_case.